меняет replace, rename utils
